LayoutTests/imported/w3c/web-platform-tests/css/css-backgrounds/animations/border-image-slice-interpolation-expected.txt


PASS CSS Transitions: property <border-image-slice> from neutral to [10%] at (-0.3) should be [23%] 
PASS CSS Transitions: property <border-image-slice> from neutral to [10%] at (0) should be [20%] 
PASS CSS Transitions: property <border-image-slice> from neutral to [10%] at (0.3) should be [17%] 
PASS CSS Transitions: property <border-image-slice> from neutral to [10%] at (0.5) should be [15%] 
PASS CSS Transitions: property <border-image-slice> from neutral to [10%] at (0.6) should be [14%] 
PASS CSS Transitions: property <border-image-slice> from neutral to [10%] at (1) should be [10%] 
PASS CSS Transitions: property <border-image-slice> from neutral to [10%] at (1.5) should be [5%] 
PASS CSS Transitions with transition: all: property <border-image-slice> from neutral to [10%] at (-0.3) should be [23%] 
PASS CSS Transitions with transition: all: property <border-image-slice> from neutral to [10%] at (0) should be [20%] 
PASS CSS Transitions with transition: all: property <border-image-slice> from neutral to [10%] at (0.3) should be [17%] 
PASS CSS Transitions with transition: all: property <border-image-slice> from neutral to [10%] at (0.5) should be [15%] 
PASS CSS Transitions with transition: all: property <border-image-slice> from neutral to [10%] at (0.6) should be [14%] 
PASS CSS Transitions with transition: all: property <border-image-slice> from neutral to [10%] at (1) should be [10%] 
PASS CSS Transitions with transition: all: property <border-image-slice> from neutral to [10%] at (1.5) should be [5%] 
PASS CSS Animations: property <border-image-slice> from neutral to [10%] at (-0.3) should be [23%] 
PASS CSS Animations: property <border-image-slice> from neutral to [10%] at (0) should be [20%] 
PASS CSS Animations: property <border-image-slice> from neutral to [10%] at (0.3) should be [17%] 
PASS CSS Animations: property <border-image-slice> from neutral to [10%] at (0.5) should be [15%] 
PASS CSS Animations: property <border-image-slice> from neutral to [10%] at (0.6) should be [14%] 
PASS CSS Animations: property <border-image-slice> from neutral to [10%] at (1) should be [10%] 
PASS CSS Animations: property <border-image-slice> from neutral to [10%] at (1.5) should be [5%] 
PASS Web Animations: property <border-image-slice> from neutral to [10%] at (-0.3) should be [23%] 
FAIL Web Animations: property <border-image-slice> from neutral to [10%] at (0) should be [20%] assert_equals: expected "20 % " but got "23 % "
FAIL Web Animations: property <border-image-slice> from neutral to [10%] at (0.3) should be [17%] assert_equals: expected "17 % " but got "19.1 % "
FAIL Web Animations: property <border-image-slice> from neutral to [10%] at (0.5) should be [15%] assert_equals: expected "15 % " but got "14.55 % "
FAIL Web Animations: property <border-image-slice> from neutral to [10%] at (0.6) should be [14%] assert_equals: expected "14 % " but got "11.82 % "
PASS Web Animations: property <border-image-slice> from neutral to [10%] at (1) should be [10%] 
FAIL Web Animations: property <border-image-slice> from neutral to [10%] at (1.5) should be [5%] assert_equals: expected "5 % " but got "10 % "
PASS CSS Transitions: property <border-image-slice> from [initial] to [10%] at (-0.3) should be [127%] 
PASS CSS Transitions: property <border-image-slice> from [initial] to [10%] at (0) should be [100%] 
PASS CSS Transitions: property <border-image-slice> from [initial] to [10%] at (0.3) should be [73%] 
PASS CSS Transitions: property <border-image-slice> from [initial] to [10%] at (0.5) should be [55%] 
PASS CSS Transitions: property <border-image-slice> from [initial] to [10%] at (0.6) should be [46%] 
PASS CSS Transitions: property <border-image-slice> from [initial] to [10%] at (1) should be [10%] 
FAIL CSS Transitions: property <border-image-slice> from [initial] to [10%] at (1.5) should be [0%] assert_equals: expected "0 % " but got "- 35 % "
PASS CSS Transitions with transition: all: property <border-image-slice> from [initial] to [10%] at (-0.3) should be [127%] 
PASS CSS Transitions with transition: all: property <border-image-slice> from [initial] to [10%] at (0) should be [100%] 
PASS CSS Transitions with transition: all: property <border-image-slice> from [initial] to [10%] at (0.3) should be [73%] 
PASS CSS Transitions with transition: all: property <border-image-slice> from [initial] to [10%] at (0.5) should be [55%] 
PASS CSS Transitions with transition: all: property <border-image-slice> from [initial] to [10%] at (0.6) should be [46%] 
PASS CSS Transitions with transition: all: property <border-image-slice> from [initial] to [10%] at (1) should be [10%] 
FAIL CSS Transitions with transition: all: property <border-image-slice> from [initial] to [10%] at (1.5) should be [0%] assert_equals: expected "0 % " but got "- 35 % "
PASS CSS Animations: property <border-image-slice> from [initial] to [10%] at (-0.3) should be [127%] 
PASS CSS Animations: property <border-image-slice> from [initial] to [10%] at (0) should be [100%] 
PASS CSS Animations: property <border-image-slice> from [initial] to [10%] at (0.3) should be [73%] 
PASS CSS Animations: property <border-image-slice> from [initial] to [10%] at (0.5) should be [55%] 
PASS CSS Animations: property <border-image-slice> from [initial] to [10%] at (0.6) should be [46%] 
PASS CSS Animations: property <border-image-slice> from [initial] to [10%] at (1) should be [10%] 
FAIL CSS Animations: property <border-image-slice> from [initial] to [10%] at (1.5) should be [0%] assert_equals: expected "0 % " but got "- 35 % "
PASS Web Animations: property <border-image-slice> from [initial] to [10%] at (-0.3) should be [127%] 
PASS Web Animations: property <border-image-slice> from [initial] to [10%] at (0) should be [100%] 
PASS Web Animations: property <border-image-slice> from [initial] to [10%] at (0.3) should be [73%] 
PASS Web Animations: property <border-image-slice> from [initial] to [10%] at (0.5) should be [55%] 
PASS Web Animations: property <border-image-slice> from [initial] to [10%] at (0.6) should be [46%] 
PASS Web Animations: property <border-image-slice> from [initial] to [10%] at (1) should be [10%] 
FAIL Web Animations: property <border-image-slice> from [initial] to [10%] at (1.5) should be [0%] assert_equals: expected "0 % " but got "- 35 % "
PASS CSS Transitions: property <border-image-slice> from [inherit] to [10%] at (-0.3) should be [62%] 
PASS CSS Transitions: property <border-image-slice> from [inherit] to [10%] at (0) should be [50%] 
PASS CSS Transitions: property <border-image-slice> from [inherit] to [10%] at (0.3) should be [38%] 
PASS CSS Transitions: property <border-image-slice> from [inherit] to [10%] at (0.5) should be [30%] 
PASS CSS Transitions: property <border-image-slice> from [inherit] to [10%] at (0.6) should be [26%] 
PASS CSS Transitions: property <border-image-slice> from [inherit] to [10%] at (1) should be [10%] 
FAIL CSS Transitions: property <border-image-slice> from [inherit] to [10%] at (1.5) should be [0%] assert_equals: expected "0 % " but got "- 10 % "
PASS CSS Transitions with transition: all: property <border-image-slice> from [inherit] to [10%] at (-0.3) should be [62%] 
PASS CSS Transitions with transition: all: property <border-image-slice> from [inherit] to [10%] at (0) should be [50%] 
PASS CSS Transitions with transition: all: property <border-image-slice> from [inherit] to [10%] at (0.3) should be [38%] 
PASS CSS Transitions with transition: all: property <border-image-slice> from [inherit] to [10%] at (0.5) should be [30%] 
PASS CSS Transitions with transition: all: property <border-image-slice> from [inherit] to [10%] at (0.6) should be [26%] 
PASS CSS Transitions with transition: all: property <border-image-slice> from [inherit] to [10%] at (1) should be [10%] 
FAIL CSS Transitions with transition: all: property <border-image-slice> from [inherit] to [10%] at (1.5) should be [0%] assert_equals: expected "0 % " but got "- 10 % "
PASS CSS Animations: property <border-image-slice> from [inherit] to [10%] at (-0.3) should be [62%] 
PASS CSS Animations: property <border-image-slice> from [inherit] to [10%] at (0) should be [50%] 
PASS CSS Animations: property <border-image-slice> from [inherit] to [10%] at (0.3) should be [38%] 
PASS CSS Animations: property <border-image-slice> from [inherit] to [10%] at (0.5) should be [30%] 
PASS CSS Animations: property <border-image-slice> from [inherit] to [10%] at (0.6) should be [26%] 
PASS CSS Animations: property <border-image-slice> from [inherit] to [10%] at (1) should be [10%] 
FAIL CSS Animations: property <border-image-slice> from [inherit] to [10%] at (1.5) should be [0%] assert_equals: expected "0 % " but got "- 10 % "
FAIL Web Animations: property <border-image-slice> from [inherit] to [10%] at (-0.3) should be [62%] assert_equals: expected "62 % " but got "23 % "
FAIL Web Animations: property <border-image-slice> from [inherit] to [10%] at (0) should be [50%] assert_equals: expected "50 % " but got "23 % "
FAIL Web Animations: property <border-image-slice> from [inherit] to [10%] at (0.3) should be [38%] assert_equals: expected "38 % " but got "19.1 % "
FAIL Web Animations: property <border-image-slice> from [inherit] to [10%] at (0.5) should be [30%] assert_equals: expected "30 % " but got "14.55 % "
FAIL Web Animations: property <border-image-slice> from [inherit] to [10%] at (0.6) should be [26%] assert_equals: expected "26 % " but got "11.82 % "
PASS Web Animations: property <border-image-slice> from [inherit] to [10%] at (1) should be [10%] 
FAIL Web Animations: property <border-image-slice> from [inherit] to [10%] at (1.5) should be [0%] assert_equals: expected "0 % " but got "10 % "
PASS CSS Transitions: property <border-image-slice> from [unset] to [10%] at (-0.3) should be [127%] 
PASS CSS Transitions: property <border-image-slice> from [unset] to [10%] at (0) should be [100%] 
PASS CSS Transitions: property <border-image-slice> from [unset] to [10%] at (0.3) should be [73%] 
PASS CSS Transitions: property <border-image-slice> from [unset] to [10%] at (0.5) should be [55%] 
PASS CSS Transitions: property <border-image-slice> from [unset] to [10%] at (0.6) should be [46%] 
PASS CSS Transitions: property <border-image-slice> from [unset] to [10%] at (1) should be [10%] 
FAIL CSS Transitions: property <border-image-slice> from [unset] to [10%] at (1.5) should be [0%] assert_equals: expected "0 % " but got "- 35 % "
PASS CSS Transitions with transition: all: property <border-image-slice> from [unset] to [10%] at (-0.3) should be [127%] 
PASS CSS Transitions with transition: all: property <border-image-slice> from [unset] to [10%] at (0) should be [100%] 
PASS CSS Transitions with transition: all: property <border-image-slice> from [unset] to [10%] at (0.3) should be [73%] 
PASS CSS Transitions with transition: all: property <border-image-slice> from [unset] to [10%] at (0.5) should be [55%] 
PASS CSS Transitions with transition: all: property <border-image-slice> from [unset] to [10%] at (0.6) should be [46%] 
PASS CSS Transitions with transition: all: property <border-image-slice> from [unset] to [10%] at (1) should be [10%] 
FAIL CSS Transitions with transition: all: property <border-image-slice> from [unset] to [10%] at (1.5) should be [0%] assert_equals: expected "0 % " but got "- 35 % "
PASS CSS Animations: property <border-image-slice> from [unset] to [10%] at (-0.3) should be [127%] 
PASS CSS Animations: property <border-image-slice> from [unset] to [10%] at (0) should be [100%] 
PASS CSS Animations: property <border-image-slice> from [unset] to [10%] at (0.3) should be [73%] 
PASS CSS Animations: property <border-image-slice> from [unset] to [10%] at (0.5) should be [55%] 
PASS CSS Animations: property <border-image-slice> from [unset] to [10%] at (0.6) should be [46%] 
PASS CSS Animations: property <border-image-slice> from [unset] to [10%] at (1) should be [10%] 
FAIL CSS Animations: property <border-image-slice> from [unset] to [10%] at (1.5) should be [0%] assert_equals: expected "0 % " but got "- 35 % "
PASS Web Animations: property <border-image-slice> from [unset] to [10%] at (-0.3) should be [127%] 
PASS Web Animations: property <border-image-slice> from [unset] to [10%] at (0) should be [100%] 
PASS Web Animations: property <border-image-slice> from [unset] to [10%] at (0.3) should be [73%] 
PASS Web Animations: property <border-image-slice> from [unset] to [10%] at (0.5) should be [55%] 
PASS Web Animations: property <border-image-slice> from [unset] to [10%] at (0.6) should be [46%] 
PASS Web Animations: property <border-image-slice> from [unset] to [10%] at (1) should be [10%] 
FAIL Web Animations: property <border-image-slice> from [unset] to [10%] at (1.5) should be [0%] assert_equals: expected "0 % " but got "- 35 % "
FAIL CSS Transitions: property <border-image-slice> from [0%] to [50%] at (-0.3) should be [0%] assert_equals: expected "0 % " but got "- 15 % "
PASS CSS Transitions: property <border-image-slice> from [0%] to [50%] at (0) should be [0%] 
PASS CSS Transitions: property <border-image-slice> from [0%] to [50%] at (0.3) should be [15%] 
PASS CSS Transitions: property <border-image-slice> from [0%] to [50%] at (0.5) should be [25%] 
PASS CSS Transitions: property <border-image-slice> from [0%] to [50%] at (0.6) should be [30%] 
PASS CSS Transitions: property <border-image-slice> from [0%] to [50%] at (1) should be [50%] 
PASS CSS Transitions: property <border-image-slice> from [0%] to [50%] at (1.5) should be [75%] 
FAIL CSS Transitions with transition: all: property <border-image-slice> from [0%] to [50%] at (-0.3) should be [0%] assert_equals: expected "0 % " but got "- 15 % "
PASS CSS Transitions with transition: all: property <border-image-slice> from [0%] to [50%] at (0) should be [0%] 
PASS CSS Transitions with transition: all: property <border-image-slice> from [0%] to [50%] at (0.3) should be [15%] 
PASS CSS Transitions with transition: all: property <border-image-slice> from [0%] to [50%] at (0.5) should be [25%] 
PASS CSS Transitions with transition: all: property <border-image-slice> from [0%] to [50%] at (0.6) should be [30%] 
PASS CSS Transitions with transition: all: property <border-image-slice> from [0%] to [50%] at (1) should be [50%] 
PASS CSS Transitions with transition: all: property <border-image-slice> from [0%] to [50%] at (1.5) should be [75%] 
FAIL CSS Animations: property <border-image-slice> from [0%] to [50%] at (-0.3) should be [0%] assert_equals: expected "0 % " but got "- 15 % "
PASS CSS Animations: property <border-image-slice> from [0%] to [50%] at (0) should be [0%] 
PASS CSS Animations: property <border-image-slice> from [0%] to [50%] at (0.3) should be [15%] 
PASS CSS Animations: property <border-image-slice> from [0%] to [50%] at (0.5) should be [25%] 
PASS CSS Animations: property <border-image-slice> from [0%] to [50%] at (0.6) should be [30%] 
PASS CSS Animations: property <border-image-slice> from [0%] to [50%] at (1) should be [50%] 
PASS CSS Animations: property <border-image-slice> from [0%] to [50%] at (1.5) should be [75%] 
FAIL Web Animations: property <border-image-slice> from [0%] to [50%] at (-0.3) should be [0%] assert_equals: expected "0 % " but got "- 15 % "
PASS Web Animations: property <border-image-slice> from [0%] to [50%] at (0) should be [0%] 
PASS Web Animations: property <border-image-slice> from [0%] to [50%] at (0.3) should be [15%] 
PASS Web Animations: property <border-image-slice> from [0%] to [50%] at (0.5) should be [25%] 
PASS Web Animations: property <border-image-slice> from [0%] to [50%] at (0.6) should be [30%] 
PASS Web Animations: property <border-image-slice> from [0%] to [50%] at (1) should be [50%] 
PASS Web Animations: property <border-image-slice> from [0%] to [50%] at (1.5) should be [75%] 
FAIL CSS Transitions: property <border-image-slice> from [0% 10% 20% 30%] to [40% 50% 60% 70%] at (-0.5) should be [0% 0% 0% 10%] assert_equals: expected "0 % 0 % 0 % 10 % " but got "- 20 % - 10 % 0 % 10 % "
PASS CSS Transitions: property <border-image-slice> from [0% 10% 20% 30%] to [40% 50% 60% 70%] at (0) should be [0% 10% 20% 30%] 
PASS CSS Transitions: property <border-image-slice> from [0% 10% 20% 30%] to [40% 50% 60% 70%] at (0.3) should be [12% 22% 32% 42%] 
PASS CSS Transitions: property <border-image-slice> from [0% 10% 20% 30%] to [40% 50% 60% 70%] at (0.5) should be [20% 30% 40% 50%] 
PASS CSS Transitions: property <border-image-slice> from [0% 10% 20% 30%] to [40% 50% 60% 70%] at (0.6) should be [24% 34% 44% 54%] 
PASS CSS Transitions: property <border-image-slice> from [0% 10% 20% 30%] to [40% 50% 60% 70%] at (1) should be [40% 50% 60% 70%] 
PASS CSS Transitions: property <border-image-slice> from [0% 10% 20% 30%] to [40% 50% 60% 70%] at (1.5) should be [60% 70% 80% 90%] 
FAIL CSS Transitions with transition: all: property <border-image-slice> from [0% 10% 20% 30%] to [40% 50% 60% 70%] at (-0.5) should be [0% 0% 0% 10%] assert_equals: expected "0 % 0 % 0 % 10 % " but got "- 20 % - 10 % 0 % 10 % "
PASS CSS Transitions with transition: all: property <border-image-slice> from [0% 10% 20% 30%] to [40% 50% 60% 70%] at (0) should be [0% 10% 20% 30%] 
PASS CSS Transitions with transition: all: property <border-image-slice> from [0% 10% 20% 30%] to [40% 50% 60% 70%] at (0.3) should be [12% 22% 32% 42%] 
PASS CSS Transitions with transition: all: property <border-image-slice> from [0% 10% 20% 30%] to [40% 50% 60% 70%] at (0.5) should be [20% 30% 40% 50%] 
PASS CSS Transitions with transition: all: property <border-image-slice> from [0% 10% 20% 30%] to [40% 50% 60% 70%] at (0.6) should be [24% 34% 44% 54%] 
PASS CSS Transitions with transition: all: property <border-image-slice> from [0% 10% 20% 30%] to [40% 50% 60% 70%] at (1) should be [40% 50% 60% 70%] 
PASS CSS Transitions with transition: all: property <border-image-slice> from [0% 10% 20% 30%] to [40% 50% 60% 70%] at (1.5) should be [60% 70% 80% 90%] 
FAIL CSS Animations: property <border-image-slice> from [0% 10% 20% 30%] to [40% 50% 60% 70%] at (-0.5) should be [0% 0% 0% 10%] assert_equals: expected "0 % 0 % 0 % 10 % " but got "- 20 % - 10 % 0 % 10 % "
PASS CSS Animations: property <border-image-slice> from [0% 10% 20% 30%] to [40% 50% 60% 70%] at (0) should be [0% 10% 20% 30%] 
PASS CSS Animations: property <border-image-slice> from [0% 10% 20% 30%] to [40% 50% 60% 70%] at (0.3) should be [12% 22% 32% 42%] 
PASS CSS Animations: property <border-image-slice> from [0% 10% 20% 30%] to [40% 50% 60% 70%] at (0.5) should be [20% 30% 40% 50%] 
PASS CSS Animations: property <border-image-slice> from [0% 10% 20% 30%] to [40% 50% 60% 70%] at (0.6) should be [24% 34% 44% 54%] 
PASS CSS Animations: property <border-image-slice> from [0% 10% 20% 30%] to [40% 50% 60% 70%] at (1) should be [40% 50% 60% 70%] 
PASS CSS Animations: property <border-image-slice> from [0% 10% 20% 30%] to [40% 50% 60% 70%] at (1.5) should be [60% 70% 80% 90%] 
FAIL Web Animations: property <border-image-slice> from [0% 10% 20% 30%] to [40% 50% 60% 70%] at (-0.5) should be [0% 0% 0% 10%] assert_equals: expected "0 % 0 % 0 % 10 % " but got "- 20 % - 10 % 0 % 10 % "
PASS Web Animations: property <border-image-slice> from [0% 10% 20% 30%] to [40% 50% 60% 70%] at (0) should be [0% 10% 20% 30%] 
PASS Web Animations: property <border-image-slice> from [0% 10% 20% 30%] to [40% 50% 60% 70%] at (0.3) should be [12% 22% 32% 42%] 
PASS Web Animations: property <border-image-slice> from [0% 10% 20% 30%] to [40% 50% 60% 70%] at (0.5) should be [20% 30% 40% 50%] 
PASS Web Animations: property <border-image-slice> from [0% 10% 20% 30%] to [40% 50% 60% 70%] at (0.6) should be [24% 34% 44% 54%] 
PASS Web Animations: property <border-image-slice> from [0% 10% 20% 30%] to [40% 50% 60% 70%] at (1) should be [40% 50% 60% 70%] 
PASS Web Animations: property <border-image-slice> from [0% 10% 20% 30%] to [40% 50% 60% 70%] at (1.5) should be [60% 70% 80% 90%] 
FAIL CSS Transitions: property <border-image-slice> from [0 10 20 30 fill] to [40 50 60 70 fill] at (-0.5) should be [0 0 0 10 fill] assert_equals: expected "0 0 0 10 fill " but got "- 20 - 10 0 10 fill "
PASS CSS Transitions: property <border-image-slice> from [0 10 20 30 fill] to [40 50 60 70 fill] at (0) should be [0 10 20 30 fill] 
PASS CSS Transitions: property <border-image-slice> from [0 10 20 30 fill] to [40 50 60 70 fill] at (0.3) should be [12 22 32 42 fill] 
PASS CSS Transitions: property <border-image-slice> from [0 10 20 30 fill] to [40 50 60 70 fill] at (0.5) should be [20 30 40 50 fill] 
PASS CSS Transitions: property <border-image-slice> from [0 10 20 30 fill] to [40 50 60 70 fill] at (0.6) should be [24 34 44 54 fill] 
PASS CSS Transitions: property <border-image-slice> from [0 10 20 30 fill] to [40 50 60 70 fill] at (1) should be [40 50 60 70 fill] 
PASS CSS Transitions: property <border-image-slice> from [0 10 20 30 fill] to [40 50 60 70 fill] at (1.5) should be [60 70 80 90 fill] 
FAIL CSS Transitions with transition: all: property <border-image-slice> from [0 10 20 30 fill] to [40 50 60 70 fill] at (-0.5) should be [0 0 0 10 fill] assert_equals: expected "0 0 0 10 fill " but got "- 20 - 10 0 10 fill "
PASS CSS Transitions with transition: all: property <border-image-slice> from [0 10 20 30 fill] to [40 50 60 70 fill] at (0) should be [0 10 20 30 fill] 
PASS CSS Transitions with transition: all: property <border-image-slice> from [0 10 20 30 fill] to [40 50 60 70 fill] at (0.3) should be [12 22 32 42 fill] 
PASS CSS Transitions with transition: all: property <border-image-slice> from [0 10 20 30 fill] to [40 50 60 70 fill] at (0.5) should be [20 30 40 50 fill] 
PASS CSS Transitions with transition: all: property <border-image-slice> from [0 10 20 30 fill] to [40 50 60 70 fill] at (0.6) should be [24 34 44 54 fill] 
PASS CSS Transitions with transition: all: property <border-image-slice> from [0 10 20 30 fill] to [40 50 60 70 fill] at (1) should be [40 50 60 70 fill] 
PASS CSS Transitions with transition: all: property <border-image-slice> from [0 10 20 30 fill] to [40 50 60 70 fill] at (1.5) should be [60 70 80 90 fill] 
FAIL CSS Animations: property <border-image-slice> from [0 10 20 30 fill] to [40 50 60 70 fill] at (-0.5) should be [0 0 0 10 fill] assert_equals: expected "0 0 0 10 fill " but got "- 20 - 10 0 10 "
FAIL CSS Animations: property <border-image-slice> from [0 10 20 30 fill] to [40 50 60 70 fill] at (0) should be [0 10 20 30 fill] assert_equals: expected "0 10 20 30 fill " but got "0 10 20 30 "
FAIL CSS Animations: property <border-image-slice> from [0 10 20 30 fill] to [40 50 60 70 fill] at (0.3) should be [12 22 32 42 fill] assert_equals: expected "12 22 32 42 fill " but got "12 22 32 42 "
FAIL CSS Animations: property <border-image-slice> from [0 10 20 30 fill] to [40 50 60 70 fill] at (0.5) should be [20 30 40 50 fill] assert_equals: expected "20 30 40 50 fill " but got "20 30 40 50 "
FAIL CSS Animations: property <border-image-slice> from [0 10 20 30 fill] to [40 50 60 70 fill] at (0.6) should be [24 34 44 54 fill] assert_equals: expected "24 34 44 54 fill " but got "24 34 44 54 "
FAIL CSS Animations: property <border-image-slice> from [0 10 20 30 fill] to [40 50 60 70 fill] at (1) should be [40 50 60 70 fill] assert_equals: expected "40 50 60 70 fill " but got "40 50 60 70 "
FAIL CSS Animations: property <border-image-slice> from [0 10 20 30 fill] to [40 50 60 70 fill] at (1.5) should be [60 70 80 90 fill] assert_equals: expected "60 70 80 90 fill " but got "60 70 80 90 "
FAIL Web Animations: property <border-image-slice> from [0 10 20 30 fill] to [40 50 60 70 fill] at (-0.5) should be [0 0 0 10 fill] assert_equals: expected "0 0 0 10 fill " but got "- 20 - 10 0 10 "
FAIL Web Animations: property <border-image-slice> from [0 10 20 30 fill] to [40 50 60 70 fill] at (0) should be [0 10 20 30 fill] assert_equals: expected "0 10 20 30 fill " but got "0 10 20 30 "
FAIL Web Animations: property <border-image-slice> from [0 10 20 30 fill] to [40 50 60 70 fill] at (0.3) should be [12 22 32 42 fill] assert_equals: expected "12 22 32 42 fill " but got "12 22 32 42 "
FAIL Web Animations: property <border-image-slice> from [0 10 20 30 fill] to [40 50 60 70 fill] at (0.5) should be [20 30 40 50 fill] assert_equals: expected "20 30 40 50 fill " but got "20 30 40 50 "
FAIL Web Animations: property <border-image-slice> from [0 10 20 30 fill] to [40 50 60 70 fill] at (0.6) should be [24 34 44 54 fill] assert_equals: expected "24 34 44 54 fill " but got "24 34 44 54 "
FAIL Web Animations: property <border-image-slice> from [0 10 20 30 fill] to [40 50 60 70 fill] at (1) should be [40 50 60 70 fill] assert_equals: expected "40 50 60 70 fill " but got "40 50 60 70 "
FAIL Web Animations: property <border-image-slice> from [0 10 20 30 fill] to [40 50 60 70 fill] at (1.5) should be [60 70 80 90 fill] assert_equals: expected "60 70 80 90 fill " but got "60 70 80 90 "
FAIL CSS Transitions: property <border-image-slice> from [0% 10 20% 30 fill] to [40% 50 60% 70 fill] at (-0.5) should be [0% 0 0% 10 fill] assert_equals: expected "0 % 0 0 % 10 fill " but got "- 20 % - 10 0 % 10 fill "
PASS CSS Transitions: property <border-image-slice> from [0% 10 20% 30 fill] to [40% 50 60% 70 fill] at (0) should be [0% 10 20% 30 fill] 
PASS CSS Transitions: property <border-image-slice> from [0% 10 20% 30 fill] to [40% 50 60% 70 fill] at (0.3) should be [12% 22 32% 42 fill] 
PASS CSS Transitions: property <border-image-slice> from [0% 10 20% 30 fill] to [40% 50 60% 70 fill] at (0.5) should be [20% 30 40% 50 fill] 
PASS CSS Transitions: property <border-image-slice> from [0% 10 20% 30 fill] to [40% 50 60% 70 fill] at (0.6) should be [24% 34 44% 54 fill] 
PASS CSS Transitions: property <border-image-slice> from [0% 10 20% 30 fill] to [40% 50 60% 70 fill] at (1) should be [40% 50 60% 70 fill] 
PASS CSS Transitions: property <border-image-slice> from [0% 10 20% 30 fill] to [40% 50 60% 70 fill] at (1.5) should be [60% 70 80% 90 fill] 
FAIL CSS Transitions with transition: all: property <border-image-slice> from [0% 10 20% 30 fill] to [40% 50 60% 70 fill] at (-0.5) should be [0% 0 0% 10 fill] assert_equals: expected "0 % 0 0 % 10 fill " but got "- 20 % - 10 0 % 10 fill "
PASS CSS Transitions with transition: all: property <border-image-slice> from [0% 10 20% 30 fill] to [40% 50 60% 70 fill] at (0) should be [0% 10 20% 30 fill] 
PASS CSS Transitions with transition: all: property <border-image-slice> from [0% 10 20% 30 fill] to [40% 50 60% 70 fill] at (0.3) should be [12% 22 32% 42 fill] 
PASS CSS Transitions with transition: all: property <border-image-slice> from [0% 10 20% 30 fill] to [40% 50 60% 70 fill] at (0.5) should be [20% 30 40% 50 fill] 
PASS CSS Transitions with transition: all: property <border-image-slice> from [0% 10 20% 30 fill] to [40% 50 60% 70 fill] at (0.6) should be [24% 34 44% 54 fill] 
PASS CSS Transitions with transition: all: property <border-image-slice> from [0% 10 20% 30 fill] to [40% 50 60% 70 fill] at (1) should be [40% 50 60% 70 fill] 
PASS CSS Transitions with transition: all: property <border-image-slice> from [0% 10 20% 30 fill] to [40% 50 60% 70 fill] at (1.5) should be [60% 70 80% 90 fill] 
FAIL CSS Animations: property <border-image-slice> from [0% 10 20% 30 fill] to [40% 50 60% 70 fill] at (-0.5) should be [0% 0 0% 10 fill] assert_equals: expected "0 % 0 0 % 10 fill " but got "- 20 % - 10 0 % 10 "
FAIL CSS Animations: property <border-image-slice> from [0% 10 20% 30 fill] to [40% 50 60% 70 fill] at (0) should be [0% 10 20% 30 fill] assert_equals: expected "0 % 10 20 % 30 fill " but got "0 % 10 20 % 30 "
FAIL CSS Animations: property <border-image-slice> from [0% 10 20% 30 fill] to [40% 50 60% 70 fill] at (0.3) should be [12% 22 32% 42 fill] assert_equals: expected "12 % 22 32 % 42 fill " but got "12 % 22 32 % 42 "
FAIL CSS Animations: property <border-image-slice> from [0% 10 20% 30 fill] to [40% 50 60% 70 fill] at (0.5) should be [20% 30 40% 50 fill] assert_equals: expected "20 % 30 40 % 50 fill " but got "20 % 30 40 % 50 "
FAIL CSS Animations: property <border-image-slice> from [0% 10 20% 30 fill] to [40% 50 60% 70 fill] at (0.6) should be [24% 34 44% 54 fill] assert_equals: expected "24 % 34 44 % 54 fill " but got "24 % 34 44 % 54 "
FAIL CSS Animations: property <border-image-slice> from [0% 10 20% 30 fill] to [40% 50 60% 70 fill] at (1) should be [40% 50 60% 70 fill] assert_equals: expected "40 % 50 60 % 70 fill " but got "40 % 50 60 % 70 "
FAIL CSS Animations: property <border-image-slice> from [0% 10 20% 30 fill] to [40% 50 60% 70 fill] at (1.5) should be [60% 70 80% 90 fill] assert_equals: expected "60 % 70 80 % 90 fill " but got "60 % 70 80 % 90 "
FAIL Web Animations: property <border-image-slice> from [0% 10 20% 30 fill] to [40% 50 60% 70 fill] at (-0.5) should be [0% 0 0% 10 fill] assert_equals: expected "0 % 0 0 % 10 fill " but got "- 20 % - 10 0 % 10 "
FAIL Web Animations: property <border-image-slice> from [0% 10 20% 30 fill] to [40% 50 60% 70 fill] at (0) should be [0% 10 20% 30 fill] assert_equals: expected "0 % 10 20 % 30 fill " but got "0 % 10 20 % 30 "
FAIL Web Animations: property <border-image-slice> from [0% 10 20% 30 fill] to [40% 50 60% 70 fill] at (0.3) should be [12% 22 32% 42 fill] assert_equals: expected "12 % 22 32 % 42 fill " but got "12 % 22 32 % 42 "
FAIL Web Animations: property <border-image-slice> from [0% 10 20% 30 fill] to [40% 50 60% 70 fill] at (0.5) should be [20% 30 40% 50 fill] assert_equals: expected "20 % 30 40 % 50 fill " but got "20 % 30 40 % 50 "
FAIL Web Animations: property <border-image-slice> from [0% 10 20% 30 fill] to [40% 50 60% 70 fill] at (0.6) should be [24% 34 44% 54 fill] assert_equals: expected "24 % 34 44 % 54 fill " but got "24 % 34 44 % 54 "
FAIL Web Animations: property <border-image-slice> from [0% 10 20% 30 fill] to [40% 50 60% 70 fill] at (1) should be [40% 50 60% 70 fill] assert_equals: expected "40 % 50 60 % 70 fill " but got "40 % 50 60 % 70 "
FAIL Web Animations: property <border-image-slice> from [0% 10 20% 30 fill] to [40% 50 60% 70 fill] at (1.5) should be [60% 70 80% 90 fill] assert_equals: expected "60 % 70 80 % 90 fill " but got "60 % 70 80 % 90 "
FAIL CSS Transitions: property <border-image-slice> from [0% fill] to [50%] at (-0.3) should be [50%] assert_equals: expected "50 % " but got "- 15 % "
FAIL CSS Transitions: property <border-image-slice> from [0% fill] to [50%] at (0) should be [50%] assert_equals: expected "50 % " but got "0 % "
FAIL CSS Transitions: property <border-image-slice> from [0% fill] to [50%] at (0.3) should be [50%] assert_equals: expected "50 % " but got "15 % "
FAIL CSS Transitions: property <border-image-slice> from [0% fill] to [50%] at (0.5) should be [50%] assert_equals: expected "50 % " but got "25 % "
FAIL CSS Transitions: property <border-image-slice> from [0% fill] to [50%] at (0.6) should be [50%] assert_equals: expected "50 % " but got "30 % "
PASS CSS Transitions: property <border-image-slice> from [0% fill] to [50%] at (1) should be [50%] 
FAIL CSS Transitions: property <border-image-slice> from [0% fill] to [50%] at (1.5) should be [50%] assert_equals: expected "50 % " but got "75 % "
FAIL CSS Transitions with transition: all: property <border-image-slice> from [0% fill] to [50%] at (-0.3) should be [50%] assert_equals: expected "50 % " but got "- 15 % "
FAIL CSS Transitions with transition: all: property <border-image-slice> from [0% fill] to [50%] at (0) should be [50%] assert_equals: expected "50 % " but got "0 % "
FAIL CSS Transitions with transition: all: property <border-image-slice> from [0% fill] to [50%] at (0.3) should be [50%] assert_equals: expected "50 % " but got "15 % "
FAIL CSS Transitions with transition: all: property <border-image-slice> from [0% fill] to [50%] at (0.5) should be [50%] assert_equals: expected "50 % " but got "25 % "
FAIL CSS Transitions with transition: all: property <border-image-slice> from [0% fill] to [50%] at (0.6) should be [50%] assert_equals: expected "50 % " but got "30 % "
PASS CSS Transitions with transition: all: property <border-image-slice> from [0% fill] to [50%] at (1) should be [50%] 
FAIL CSS Transitions with transition: all: property <border-image-slice> from [0% fill] to [50%] at (1.5) should be [50%] assert_equals: expected "50 % " but got "75 % "
FAIL CSS Animations: property <border-image-slice> from [0% fill] to [50%] at (-0.3) should be [0% fill] assert_equals: expected "0 % fill " but got "- 15 % "
FAIL CSS Animations: property <border-image-slice> from [0% fill] to [50%] at (0) should be [0% fill] assert_equals: expected "0 % fill " but got "0 % "
FAIL CSS Animations: property <border-image-slice> from [0% fill] to [50%] at (0.3) should be [0% fill] assert_equals: expected "0 % fill " but got "15 % "
FAIL CSS Animations: property <border-image-slice> from [0% fill] to [50%] at (0.5) should be [50%] assert_equals: expected "50 % " but got "25 % "
FAIL CSS Animations: property <border-image-slice> from [0% fill] to [50%] at (0.6) should be [50%] assert_equals: expected "50 % " but got "30 % "
PASS CSS Animations: property <border-image-slice> from [0% fill] to [50%] at (1) should be [50%] 
FAIL CSS Animations: property <border-image-slice> from [0% fill] to [50%] at (1.5) should be [50%] assert_equals: expected "50 % " but got "75 % "
FAIL Web Animations: property <border-image-slice> from [0% fill] to [50%] at (-0.3) should be [0% fill] assert_equals: expected "0 % fill " but got "- 15 % "
FAIL Web Animations: property <border-image-slice> from [0% fill] to [50%] at (0) should be [0% fill] assert_equals: expected "0 % fill " but got "0 % "
FAIL Web Animations: property <border-image-slice> from [0% fill] to [50%] at (0.3) should be [0% fill] assert_equals: expected "0 % fill " but got "15 % "
FAIL Web Animations: property <border-image-slice> from [0% fill] to [50%] at (0.5) should be [50%] assert_equals: expected "50 % " but got "25 % "
FAIL Web Animations: property <border-image-slice> from [0% fill] to [50%] at (0.6) should be [50%] assert_equals: expected "50 % " but got "30 % "
PASS Web Animations: property <border-image-slice> from [0% fill] to [50%] at (1) should be [50%] 
FAIL Web Animations: property <border-image-slice> from [0% fill] to [50%] at (1.5) should be [50%] assert_equals: expected "50 % " but got "75 % "
FAIL CSS Transitions: property <border-image-slice> from [50%] to [100] at (-0.3) should be [100] assert_equals: expected "100 " but got "50 % "
FAIL CSS Transitions: property <border-image-slice> from [50%] to [100] at (0) should be [100] assert_equals: expected "100 " but got "50 % "
FAIL CSS Transitions: property <border-image-slice> from [50%] to [100] at (0.3) should be [100] assert_equals: expected "100 " but got "0 "
FAIL CSS Transitions: property <border-image-slice> from [50%] to [100] at (0.5) should be [100] assert_equals: expected "100 " but got "0 "
FAIL CSS Transitions: property <border-image-slice> from [50%] to [100] at (0.6) should be [100] assert_equals: expected "100 " but got "0 "
PASS CSS Transitions: property <border-image-slice> from [50%] to [100] at (1) should be [100] 
PASS CSS Transitions: property <border-image-slice> from [50%] to [100] at (1.5) should be [100] 
FAIL CSS Transitions with transition: all: property <border-image-slice> from [50%] to [100] at (-0.3) should be [100] assert_equals: expected "100 " but got "50 % "
FAIL CSS Transitions with transition: all: property <border-image-slice> from [50%] to [100] at (0) should be [100] assert_equals: expected "100 " but got "50 % "
FAIL CSS Transitions with transition: all: property <border-image-slice> from [50%] to [100] at (0.3) should be [100] assert_equals: expected "100 " but got "0 "
FAIL CSS Transitions with transition: all: property <border-image-slice> from [50%] to [100] at (0.5) should be [100] assert_equals: expected "100 " but got "0 "
FAIL CSS Transitions with transition: all: property <border-image-slice> from [50%] to [100] at (0.6) should be [100] assert_equals: expected "100 " but got "0 "
PASS CSS Transitions with transition: all: property <border-image-slice> from [50%] to [100] at (1) should be [100] 
PASS CSS Transitions with transition: all: property <border-image-slice> from [50%] to [100] at (1.5) should be [100] 
PASS CSS Animations: property <border-image-slice> from [50%] to [100] at (-0.3) should be [50%] 
PASS CSS Animations: property <border-image-slice> from [50%] to [100] at (0) should be [50%] 
FAIL CSS Animations: property <border-image-slice> from [50%] to [100] at (0.3) should be [50%] assert_equals: expected "50 % " but got "0 "
FAIL CSS Animations: property <border-image-slice> from [50%] to [100] at (0.5) should be [100] assert_equals: expected "100 " but got "0 "
FAIL CSS Animations: property <border-image-slice> from [50%] to [100] at (0.6) should be [100] assert_equals: expected "100 " but got "0 "
PASS CSS Animations: property <border-image-slice> from [50%] to [100] at (1) should be [100] 
PASS CSS Animations: property <border-image-slice> from [50%] to [100] at (1.5) should be [100] 
PASS Web Animations: property <border-image-slice> from [50%] to [100] at (-0.3) should be [50%] 
PASS Web Animations: property <border-image-slice> from [50%] to [100] at (0) should be [50%] 
FAIL Web Animations: property <border-image-slice> from [50%] to [100] at (0.3) should be [50%] assert_equals: expected "50 % " but got "0 "
FAIL Web Animations: property <border-image-slice> from [50%] to [100] at (0.5) should be [100] assert_equals: expected "100 " but got "0 "
FAIL Web Animations: property <border-image-slice> from [50%] to [100] at (0.6) should be [100] assert_equals: expected "100 " but got "0 "
PASS Web Animations: property <border-image-slice> from [50%] to [100] at (1) should be [100] 
PASS Web Animations: property <border-image-slice> from [50%] to [100] at (1.5) should be [100] 
FAIL CSS Transitions: property <border-image-slice> from [50% fill] to [100 fill] at (-0.3) should be [100 fill] assert_equals: expected "100 fill " but got "50 % fill "
FAIL CSS Transitions: property <border-image-slice> from [50% fill] to [100 fill] at (0) should be [100 fill] assert_equals: expected "100 fill " but got "50 % fill "
FAIL CSS Transitions: property <border-image-slice> from [50% fill] to [100 fill] at (0.3) should be [100 fill] assert_equals: expected "100 fill " but got "0 fill "
FAIL CSS Transitions: property <border-image-slice> from [50% fill] to [100 fill] at (0.5) should be [100 fill] assert_equals: expected "100 fill " but got "0 fill "
FAIL CSS Transitions: property <border-image-slice> from [50% fill] to [100 fill] at (0.6) should be [100 fill] assert_equals: expected "100 fill " but got "0 fill "
PASS CSS Transitions: property <border-image-slice> from [50% fill] to [100 fill] at (1) should be [100 fill] 
PASS CSS Transitions: property <border-image-slice> from [50% fill] to [100 fill] at (1.5) should be [100 fill] 
FAIL CSS Transitions with transition: all: property <border-image-slice> from [50% fill] to [100 fill] at (-0.3) should be [100 fill] assert_equals: expected "100 fill " but got "50 % fill "
FAIL CSS Transitions with transition: all: property <border-image-slice> from [50% fill] to [100 fill] at (0) should be [100 fill] assert_equals: expected "100 fill " but got "50 % fill "
FAIL CSS Transitions with transition: all: property <border-image-slice> from [50% fill] to [100 fill] at (0.3) should be [100 fill] assert_equals: expected "100 fill " but got "0 fill "
FAIL CSS Transitions with transition: all: property <border-image-slice> from [50% fill] to [100 fill] at (0.5) should be [100 fill] assert_equals: expected "100 fill " but got "0 fill "
FAIL CSS Transitions with transition: all: property <border-image-slice> from [50% fill] to [100 fill] at (0.6) should be [100 fill] assert_equals: expected "100 fill " but got "0 fill "
PASS CSS Transitions with transition: all: property <border-image-slice> from [50% fill] to [100 fill] at (1) should be [100 fill] 
PASS CSS Transitions with transition: all: property <border-image-slice> from [50% fill] to [100 fill] at (1.5) should be [100 fill] 
FAIL CSS Animations: property <border-image-slice> from [50% fill] to [100 fill] at (-0.3) should be [50% fill] assert_equals: expected "50 % fill " but got "50 % "
FAIL CSS Animations: property <border-image-slice> from [50% fill] to [100 fill] at (0) should be [50% fill] assert_equals: expected "50 % fill " but got "50 % "
FAIL CSS Animations: property <border-image-slice> from [50% fill] to [100 fill] at (0.3) should be [50% fill] assert_equals: expected "50 % fill " but got "0 "
FAIL CSS Animations: property <border-image-slice> from [50% fill] to [100 fill] at (0.5) should be [100 fill] assert_equals: expected "100 fill " but got "0 "
FAIL CSS Animations: property <border-image-slice> from [50% fill] to [100 fill] at (0.6) should be [100 fill] assert_equals: expected "100 fill " but got "0 "
FAIL CSS Animations: property <border-image-slice> from [50% fill] to [100 fill] at (1) should be [100 fill] assert_equals: expected "100 fill " but got "100 "
FAIL CSS Animations: property <border-image-slice> from [50% fill] to [100 fill] at (1.5) should be [100 fill] assert_equals: expected "100 fill " but got "100 "
FAIL Web Animations: property <border-image-slice> from [50% fill] to [100 fill] at (-0.3) should be [50% fill] assert_equals: expected "50 % fill " but got "50 % "
FAIL Web Animations: property <border-image-slice> from [50% fill] to [100 fill] at (0) should be [50% fill] assert_equals: expected "50 % fill " but got "50 % "
FAIL Web Animations: property <border-image-slice> from [50% fill] to [100 fill] at (0.3) should be [50% fill] assert_equals: expected "50 % fill " but got "0 "
FAIL Web Animations: property <border-image-slice> from [50% fill] to [100 fill] at (0.5) should be [100 fill] assert_equals: expected "100 fill " but got "0 "
FAIL Web Animations: property <border-image-slice> from [50% fill] to [100 fill] at (0.6) should be [100 fill] assert_equals: expected "100 fill " but got "0 "
FAIL Web Animations: property <border-image-slice> from [50% fill] to [100 fill] at (1) should be [100 fill] assert_equals: expected "100 fill " but got "100 "
FAIL Web Animations: property <border-image-slice> from [50% fill] to [100 fill] at (1.5) should be [100 fill] assert_equals: expected "100 fill " but got "100 "
FAIL CSS Transitions: property <border-image-slice> from [0% 10 20% 30 fill] to [40% 50 60% 70] at (-0.3) should be [40% 50 60% 70] assert_equals: expected "40 % 50 60 % 70 " but got "- 12 % - 2 8 % 18 "
FAIL CSS Transitions: property <border-image-slice> from [0% 10 20% 30 fill] to [40% 50 60% 70] at (0) should be [40% 50 60% 70] assert_equals: expected "40 % 50 60 % 70 " but got "0 % 10 20 % 30 "
FAIL CSS Transitions: property <border-image-slice> from [0% 10 20% 30 fill] to [40% 50 60% 70] at (0.3) should be [40% 50 60% 70] assert_equals: expected "40 % 50 60 % 70 " but got "12 % 22 32 % 42 "
FAIL CSS Transitions: property <border-image-slice> from [0% 10 20% 30 fill] to [40% 50 60% 70] at (0.5) should be [40% 50 60% 70] assert_equals: expected "40 % 50 60 % 70 " but got "20 % 30 40 % 50 "
FAIL CSS Transitions: property <border-image-slice> from [0% 10 20% 30 fill] to [40% 50 60% 70] at (0.6) should be [40% 50 60% 70] assert_equals: expected "40 % 50 60 % 70 " but got "24 % 34 44 % 54 "
PASS CSS Transitions: property <border-image-slice> from [0% 10 20% 30 fill] to [40% 50 60% 70] at (1) should be [40% 50 60% 70] 
FAIL CSS Transitions: property <border-image-slice> from [0% 10 20% 30 fill] to [40% 50 60% 70] at (1.5) should be [40% 50 60% 70] assert_equals: expected "40 % 50 60 % 70 " but got "60 % 70 80 % 90 "
FAIL CSS Transitions with transition: all: property <border-image-slice> from [0% 10 20% 30 fill] to [40% 50 60% 70] at (-0.3) should be [40% 50 60% 70] assert_equals: expected "40 % 50 60 % 70 " but got "- 12 % - 2 8 % 18 "
FAIL CSS Transitions with transition: all: property <border-image-slice> from [0% 10 20% 30 fill] to [40% 50 60% 70] at (0) should be [40% 50 60% 70] assert_equals: expected "40 % 50 60 % 70 " but got "0 % 10 20 % 30 "
FAIL CSS Transitions with transition: all: property <border-image-slice> from [0% 10 20% 30 fill] to [40% 50 60% 70] at (0.3) should be [40% 50 60% 70] assert_equals: expected "40 % 50 60 % 70 " but got "12 % 22 32 % 42 "
FAIL CSS Transitions with transition: all: property <border-image-slice> from [0% 10 20% 30 fill] to [40% 50 60% 70] at (0.5) should be [40% 50 60% 70] assert_equals: expected "40 % 50 60 % 70 " but got "20 % 30 40 % 50 "
FAIL CSS Transitions with transition: all: property <border-image-slice> from [0% 10 20% 30 fill] to [40% 50 60% 70] at (0.6) should be [40% 50 60% 70] assert_equals: expected "40 % 50 60 % 70 " but got "24 % 34 44 % 54 "
PASS CSS Transitions with transition: all: property <border-image-slice> from [0% 10 20% 30 fill] to [40% 50 60% 70] at (1) should be [40% 50 60% 70] 
FAIL CSS Transitions with transition: all: property <border-image-slice> from [0% 10 20% 30 fill] to [40% 50 60% 70] at (1.5) should be [40% 50 60% 70] assert_equals: expected "40 % 50 60 % 70 " but got "60 % 70 80 % 90 "
FAIL CSS Animations: property <border-image-slice> from [0% 10 20% 30 fill] to [40% 50 60% 70] at (-0.3) should be [0% 10 20% 30 fill] assert_equals: expected "0 % 10 20 % 30 fill " but got "- 12 % - 2 8 % 18 "
FAIL CSS Animations: property <border-image-slice> from [0% 10 20% 30 fill] to [40% 50 60% 70] at (0) should be [0% 10 20% 30 fill] assert_equals: expected "0 % 10 20 % 30 fill " but got "0 % 10 20 % 30 "
FAIL CSS Animations: property <border-image-slice> from [0% 10 20% 30 fill] to [40% 50 60% 70] at (0.3) should be [0% 10 20% 30 fill] assert_equals: expected "0 % 10 20 % 30 fill " but got "12 % 22 32 % 42 "
FAIL CSS Animations: property <border-image-slice> from [0% 10 20% 30 fill] to [40% 50 60% 70] at (0.5) should be [40% 50 60% 70] assert_equals: expected "40 % 50 60 % 70 " but got "20 % 30 40 % 50 "
FAIL CSS Animations: property <border-image-slice> from [0% 10 20% 30 fill] to [40% 50 60% 70] at (0.6) should be [40% 50 60% 70] assert_equals: expected "40 % 50 60 % 70 " but got "24 % 34 44 % 54 "
PASS CSS Animations: property <border-image-slice> from [0% 10 20% 30 fill] to [40% 50 60% 70] at (1) should be [40% 50 60% 70] 
FAIL CSS Animations: property <border-image-slice> from [0% 10 20% 30 fill] to [40% 50 60% 70] at (1.5) should be [40% 50 60% 70] assert_equals: expected "40 % 50 60 % 70 " but got "60 % 70 80 % 90 "
FAIL Web Animations: property <border-image-slice> from [0% 10 20% 30 fill] to [40% 50 60% 70] at (-0.3) should be [0% 10 20% 30 fill] assert_equals: expected "0 % 10 20 % 30 fill " but got "- 12 % - 2 8 % 18 "
FAIL Web Animations: property <border-image-slice> from [0% 10 20% 30 fill] to [40% 50 60% 70] at (0) should be [0% 10 20% 30 fill] assert_equals: expected "0 % 10 20 % 30 fill " but got "0 % 10 20 % 30 "
FAIL Web Animations: property <border-image-slice> from [0% 10 20% 30 fill] to [40% 50 60% 70] at (0.3) should be [0% 10 20% 30 fill] assert_equals: expected "0 % 10 20 % 30 fill " but got "12 % 22 32 % 42 "
FAIL Web Animations: property <border-image-slice> from [0% 10 20% 30 fill] to [40% 50 60% 70] at (0.5) should be [40% 50 60% 70] assert_equals: expected "40 % 50 60 % 70 " but got "20 % 30 40 % 50 "
FAIL Web Animations: property <border-image-slice> from [0% 10 20% 30 fill] to [40% 50 60% 70] at (0.6) should be [40% 50 60% 70] assert_equals: expected "40 % 50 60 % 70 " but got "24 % 34 44 % 54 "
PASS Web Animations: property <border-image-slice> from [0% 10 20% 30 fill] to [40% 50 60% 70] at (1) should be [40% 50 60% 70] 
FAIL Web Animations: property <border-image-slice> from [0% 10 20% 30 fill] to [40% 50 60% 70] at (1.5) should be [40% 50 60% 70] assert_equals: expected "40 % 50 60 % 70 " but got "60 % 70 80 % 90 "
FAIL CSS Transitions: property <border-image-slice> from [0% 10 20 30 fill] to [40 50 60% 70] at (-0.3) should be [40 50 60% 70] assert_equals: expected "40 50 60 % 70 " but got "- 12 - 2 20 18 "
FAIL CSS Transitions: property <border-image-slice> from [0% 10 20 30 fill] to [40 50 60% 70] at (0) should be [40 50 60% 70] assert_equals: expected "40 50 60 % 70 " but got "0 10 20 30 "
FAIL CSS Transitions: property <border-image-slice> from [0% 10 20 30 fill] to [40 50 60% 70] at (0.3) should be [40 50 60% 70] assert_equals: expected "40 50 60 % 70 " but got "12 22 0 42 "
FAIL CSS Transitions: property <border-image-slice> from [0% 10 20 30 fill] to [40 50 60% 70] at (0.5) should be [40 50 60% 70] assert_equals: expected "40 50 60 % 70 " but got "20 30 0 50 "
FAIL CSS Transitions: property <border-image-slice> from [0% 10 20 30 fill] to [40 50 60% 70] at (0.6) should be [40 50 60% 70] assert_equals: expected "40 50 60 % 70 " but got "24 34 0 54 "
PASS CSS Transitions: property <border-image-slice> from [0% 10 20 30 fill] to [40 50 60% 70] at (1) should be [40 50 60% 70] 
FAIL CSS Transitions: property <border-image-slice> from [0% 10 20 30 fill] to [40 50 60% 70] at (1.5) should be [40 50 60% 70] assert_equals: expected "40 50 60 % 70 " but got "60 70 60 % 90 "
FAIL CSS Transitions with transition: all: property <border-image-slice> from [0% 10 20 30 fill] to [40 50 60% 70] at (-0.3) should be [40 50 60% 70] assert_equals: expected "40 50 60 % 70 " but got "- 12 - 2 20 18 "
FAIL CSS Transitions with transition: all: property <border-image-slice> from [0% 10 20 30 fill] to [40 50 60% 70] at (0) should be [40 50 60% 70] assert_equals: expected "40 50 60 % 70 " but got "0 10 20 30 "
FAIL CSS Transitions with transition: all: property <border-image-slice> from [0% 10 20 30 fill] to [40 50 60% 70] at (0.3) should be [40 50 60% 70] assert_equals: expected "40 50 60 % 70 " but got "12 22 0 42 "
FAIL CSS Transitions with transition: all: property <border-image-slice> from [0% 10 20 30 fill] to [40 50 60% 70] at (0.5) should be [40 50 60% 70] assert_equals: expected "40 50 60 % 70 " but got "20 30 0 50 "
FAIL CSS Transitions with transition: all: property <border-image-slice> from [0% 10 20 30 fill] to [40 50 60% 70] at (0.6) should be [40 50 60% 70] assert_equals: expected "40 50 60 % 70 " but got "24 34 0 54 "
PASS CSS Transitions with transition: all: property <border-image-slice> from [0% 10 20 30 fill] to [40 50 60% 70] at (1) should be [40 50 60% 70] 
FAIL CSS Transitions with transition: all: property <border-image-slice> from [0% 10 20 30 fill] to [40 50 60% 70] at (1.5) should be [40 50 60% 70] assert_equals: expected "40 50 60 % 70 " but got "60 70 60 % 90 "
FAIL CSS Animations: property <border-image-slice> from [0% 10 20 30 fill] to [40 50 60% 70] at (-0.3) should be [0% 10 20 30 fill] assert_equals: expected "0 % 10 20 30 fill " but got "- 12 - 2 20 18 "
FAIL CSS Animations: property <border-image-slice> from [0% 10 20 30 fill] to [40 50 60% 70] at (0) should be [0% 10 20 30 fill] assert_equals: expected "0 % 10 20 30 fill " but got "0 10 20 30 "
FAIL CSS Animations: property <border-image-slice> from [0% 10 20 30 fill] to [40 50 60% 70] at (0.3) should be [0% 10 20 30 fill] assert_equals: expected "0 % 10 20 30 fill " but got "12 22 0 42 "
FAIL CSS Animations: property <border-image-slice> from [0% 10 20 30 fill] to [40 50 60% 70] at (0.5) should be [40 50 60% 70] assert_equals: expected "40 50 60 % 70 " but got "20 30 0 50 "
FAIL CSS Animations: property <border-image-slice> from [0% 10 20 30 fill] to [40 50 60% 70] at (0.6) should be [40 50 60% 70] assert_equals: expected "40 50 60 % 70 " but got "24 34 0 54 "
PASS CSS Animations: property <border-image-slice> from [0% 10 20 30 fill] to [40 50 60% 70] at (1) should be [40 50 60% 70] 
FAIL CSS Animations: property <border-image-slice> from [0% 10 20 30 fill] to [40 50 60% 70] at (1.5) should be [40 50 60% 70] assert_equals: expected "40 50 60 % 70 " but got "60 70 60 % 90 "
FAIL Web Animations: property <border-image-slice> from [0% 10 20 30 fill] to [40 50 60% 70] at (-0.3) should be [0% 10 20 30 fill] assert_equals: expected "0 % 10 20 30 fill " but got "- 12 - 2 20 18 "
FAIL Web Animations: property <border-image-slice> from [0% 10 20 30 fill] to [40 50 60% 70] at (0) should be [0% 10 20 30 fill] assert_equals: expected "0 % 10 20 30 fill " but got "0 10 20 30 "
FAIL Web Animations: property <border-image-slice> from [0% 10 20 30 fill] to [40 50 60% 70] at (0.3) should be [0% 10 20 30 fill] assert_equals: expected "0 % 10 20 30 fill " but got "12 22 0 42 "
FAIL Web Animations: property <border-image-slice> from [0% 10 20 30 fill] to [40 50 60% 70] at (0.5) should be [40 50 60% 70] assert_equals: expected "40 50 60 % 70 " but got "20 30 0 50 "
FAIL Web Animations: property <border-image-slice> from [0% 10 20 30 fill] to [40 50 60% 70] at (0.6) should be [40 50 60% 70] assert_equals: expected "40 50 60 % 70 " but got "24 34 0 54 "
PASS Web Animations: property <border-image-slice> from [0% 10 20 30 fill] to [40 50 60% 70] at (1) should be [40 50 60% 70] 
FAIL Web Animations: property <border-image-slice> from [0% 10 20 30 fill] to [40 50 60% 70] at (1.5) should be [40 50 60% 70] assert_equals: expected "40 50 60 % 70 " but got "60 70 60 % 90 "
